requirements.txt
PyYAML>=6.0.1
python-dotenv>=1.0.1
pydub>=0.25.1
tqdm>=4.66.0
tenacity>=8.2.3
pysrt>=1.1.2
# ASR (optional but recommended):
faster-whisper>=1.0.3
# ElevenLabs official SDK:
elevenlabs>=1.0.0
argostranslate>=1.9.6
gradio>=4.0.0
requests>=2.31.0
soundfile>=0.12.1
